fix strip_screen_size

+++ src/common/strip.c
@@ -6,7 +6,7 @@
 /*   By: svogrig <svogrig@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/26 17:59:44 by svogrig           #+#    #+#             */
-/*   Updated: 2024/12/26 23:26:38 by svogrig          ###   ########.fr       */
+/*   Updated: 2025/01/07 21:10:32 by svogrig          ###   ########.fr       */
 /*                                                                            */
 /******************************************************************************/
 
@@ -14,14 +14,14 @@
 
 int	strip_screen_size(int screen_scale, double distance)
 {
-	int size;
+	float size;
 
 	if (distance == 0)
 		return (INT_MAX);
 	size = screen_scale / distance;
 	if (size > INT_MAX)
 		return (INT_MAX);
-	return (size);
+	return ((int)size);
 }
 
 void	strip_add_limit(t_strip *strip, int size)
